pyproject.toml
[project]
name = "cipher"
version = "0.1.0"
description = "Cipher — automated crypto futures trading bot"
requires-python = ">=3.12"
dependencies = [
    "fastapi==0.129.1",
    "uvicorn==0.41.0",
    "python-binance==1.0.35",
    "pydantic==2.12.5",
    "python-dotenv==1.2.1",
    "pytest==9.0.2",
    "httpx==0.28.1",
    "requests==2.32.5",
    "pandas>=2.0",
    "numpy>=1.26",
    "pytest-asyncio>=1.3.0",
]

[build-system]
requires = ["hatchling>=1.25.0"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
asyncio_mode = "auto"

[tool.uv]
package = false
